+++ library/observables-connect/src/extensions/index.ts
@@ -4,6 +4,5 @@ SPDX-License-Identifier: GPL-3.0-only */
 export * from './connect'
 export * from './discover'
 export * from './enable'
-export * from './local'
-export * from './observables'
-export * from './state'
+export * from './init'
+export * from './reconnect'
