+++ library/observables-connect/src/index.ts
@@ -1,6 +1,5 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-export * from './get'
-export * from './unsubs'
+export * from './observables'
 export * from './util'
