+++ library/observables-connect/src/local.ts
@@ -2,7 +2,7 @@
 SPDX-License-Identifier: GPL-3.0-only */
 
 import { localStorageOrDefault } from '@w3ux/utils'
-import { activeExtensionsKey } from '../consts'
+import { activeExtensionsKey } from './consts'
 
 // Check if an extension exists in local storage
 export const isExtensionLocal = (id: string): boolean => {
