+++ library/observables-connect/src/observables.ts
@@ -1,7 +1,7 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import type { ExtensionsStatus } from '@w3ux/types'
+import type { ExtensionsStatus, ImportedAccount, Sync } from '@w3ux/types'
 import { BehaviorSubject } from 'rxjs'
 
 // Discovered extensions along with their status
@@ -15,3 +15,11 @@ export const gettingExtensions$ = _gettingExtensions.asObservable()
 // Extensions that have successfully connected
 export const _initialisedExtensions = new BehaviorSubject<string[]>([])
 export const initialisedExtensions$ = _initialisedExtensions.asObservable()
+
+// Sync status of reconnecting to previously enabled extensions
+export const _reconnectSync = new BehaviorSubject<Sync>('unsynced')
+export const reconnectSync$ = _reconnectSync.asObservable()
+
+// Imported extension accounts
+export const _extensionAccounts = new BehaviorSubject<ImportedAccount[]>([])
+export const extensionAccounts$ = _extensionAccounts.asObservable()
